Move away setUp/tearDown logic from AwsS3Test to a dedicated trait
It would ease writing functional tests for people who wants to extends Gaufrette behaviors.
For instance, Gaufrette/extras would benefit this change.
