+++ tests/Gaufrette/Functional/Adapter/AwsS3Test.php
@@ -11,65 +11,10 @@ use Gaufrette\Filesystem;
  */
 class AwsS3Test extends \PHPUnit_Framework_TestCase
 {
+    use AwsS3SetUpTearDownTrait;
+
     /** @var int */
     static private $SDK_VERSION;
-
-    /** @var string */
-    private $bucket;
-
-    /** @var S3Client */
-    private $client;
-
-    public function setUp()
-    {
-        $key = getenv('AWS_KEY');
-        $secret = getenv('AWS_SECRET');
-
-        if (empty($key) || empty($secret)) {
-            $this->markTestSkipped();
-        }
-
-        if (self::$SDK_VERSION === null) {
-            self::$SDK_VERSION = method_exists(S3Client::class, 'getArguments') ? 3 : 2;
-        }
-
-        $this->bucket = uniqid(getenv('AWS_BUCKET'));
-
-        if (self::$SDK_VERSION === 3) {
-            // New way of instantiating S3Client for aws-sdk-php v3
-            $this->client = new S3Client([
-                'region' => 'eu-west-1',
-                'version' => 'latest',
-                'credentials' => [
-                    'key' => $key,
-                    'secret' => $secret,
-                ],
-            ]);
-        } else {
-            $this->client = S3Client::factory([
-                'region' => 'eu-west-1',
-                'version' => '2006-03-01',
-                'key' => $key,
-                'secret' => $secret,
-            ]);
-        }
-    }
-
-    public function tearDown()
-    {
-        if ($this->client === null || !$this->client->doesBucketExist($this->bucket)) {
-            return;
-        }
-
-        $result = $this->client->listObjects(['Bucket' => $this->bucket]);
-        $staleObjects = $result->get('Contents');
-
-        foreach ($staleObjects as $staleObject) {
-            $this->client->deleteObject(['Bucket' => $this->bucket, 'Key' => $staleObject['Key']]);
-        }
-
-        $this->client->deleteBucket(['Bucket' => $this->bucket]);
-    }
 
     private function getFilesystem(array $adapterOptions = [])
     {
